moves decision logic into Boardstate class
